+++ test/lightning_web/live/workflow_live_test.exs
@@ -1,0 +1,93 @@
+defmodule LightningWeb.WorkflowLiveTest do
+  use LightningWeb.ConnCase, async: true
+
+  import Phoenix.LiveViewTest
+
+  setup :register_and_log_in_user
+  setup :create_project_for_current_user
+
+  import Lightning.JobsFixtures
+
+  describe "show" do
+    setup %{project: project} do
+      %{job: job_fixture(project_id: project.id)}
+    end
+
+    test "renders the workflow diagram", %{
+      conn: conn,
+      project: project
+    } do
+      {:ok, view, html} =
+        live(conn, Routes.project_workflow_path(conn, :show, project.id))
+
+      assert html =~ project.name
+
+      expected_encoded_project_space =
+        Lightning.Workflows.get_workflows_for(project)
+        |> Lightning.Workflows.to_project_space()
+        |> Jason.encode!()
+        |> Base.encode64()
+
+      assert view
+             |> element("div#hook-#{project.id}[phx-update=ignore]")
+             |> render() =~ expected_encoded_project_space
+    end
+  end
+
+  describe "edit_job" do
+    setup %{project: project} do
+      %{job: job_fixture(project_id: project.id)}
+    end
+
+    test "renders the job inspector", %{
+      conn: conn,
+      project: project,
+      job: job
+    } do
+      {:ok, view, html} =
+        live(
+          conn,
+          Routes.project_workflow_path(conn, :edit_job, project.id, job.id)
+        )
+
+      assert html =~ project.name
+
+      assert has_element?(view, "#job-#{job.id}")
+    end
+  end
+
+  describe "edit_workflow" do
+    setup %{project: project} do
+      %{job: job_fixture(project_id: project.id)}
+    end
+
+    test "renders the workflow inspector", %{
+      conn: conn,
+      project: project,
+      job: job
+    } do
+      {:ok, view, html} =
+        live(
+          conn,
+          Routes.project_workflow_path(
+            conn,
+            :edit_workflow,
+            project.id,
+            job.workflow_id
+          )
+        )
+
+      assert html =~ project.name
+
+      assert has_element?(view, "#workflow-#{job.workflow_id}")
+
+      assert view
+             |> form("#workflow-form", workflow: %{name: "my workflow"})
+             |> render_change()
+
+      view |> form("#workflow-form") |> render_submit()
+
+      assert_patch(view, Routes.project_workflow_path(conn, :show, project.id))
+    end
+  end
+end